refactor(): moved code into src , moved old code to src/old
