+++ src/parseStlAsStreamNoWorker.js
