+++ src/parseStlAsStreamWorker.js
