+++ src/parsers/stl/parseHelpers.js
