+++ src/parsers/stl/parseStream.js
