+++ src/parsers/stl/parseStreamAlt.js
