+++ src/parsers/stl/parseStreamAlt2.js
