+++ src/parsers/stl/utils.js
