+++ src/readFileBasic.js
