+++ src/workers/spawners/streamWorkerSpawner.js
