+++ src/workers/spawners/streamWorkerSpawner2.js
