+++ src/workers/workers/stlStreamWorker.js
