+++ src/workers/workers/stlStreamWorker2.js
