+++ src/workers/workers/stlStreamWorker2.src.js
